Improve of overlay for elements detect confirmation and removal of old implementation files from previous logic of detection
