+++ src/content/actions/MapperActions.ts
@@ -1,4 +1,4 @@
-import { actionCorrectElementSelection, actionDetectChatbot, actionEndSucessfulVerification, actionIdentifyChatbotSelectors, actionSelectMicrophone, actionSetStoredSelectors, actionStartSelection, actionStartVerification, startPageChatbotProcedure } from "./ActionsDetection";
+import { actionCorrectElementSelection, actionDetectChatbot, actionEndSucessfulVerification, actionIdentifyChatbotSelectors,  actionSetStoredSelectors, actionStartVerification, startPageChatbotProcedure } from "./ActionsDetection";
 import { actionEndEvaluation, actionEvaluateACT, actionEvaluateCUI } from "./ActionsEvaluation";
 import { actionLLMInteraction, actionStartVoiceInput, actionTypeMessages } from "./ActionsInteraction";
 
@@ -16,14 +16,12 @@ const caseInteractionHandlers: Record<string, (data:IChromeRequest) => void | Pr
 
 // handlers for the detection of chatbot actions
 const caseDetectionHandlers: Record<string, (data:IChromeRequest) => void | Promise<any>> = {
-    startSelection: actionStartSelection,  
     detectChatbot: actionDetectChatbot,
     pageChatbotProcedure: startPageChatbotProcedure,
     identifySelectors: actionIdentifyChatbotSelectors, 
     startVerification: actionStartVerification,  
     endSucessfulVerification: actionEndSucessfulVerification , 
     correctElementSelection: actionCorrectElementSelection, 
-    startMicSelection: actionSelectMicrophone,
     setStoredSelectors: actionSetStoredSelectors,
 }
